Move hasNoAIA
